+++ lizard_progress/tasks.py
@@ -10,6 +10,7 @@ from celery.task import task
 
 from lizard_progress import process_uploaded_file
 from lizard_progress import exports
+from lizard_progress.util import shapevac
 
 
 @task
@@ -28,3 +29,9 @@ def process_uploaded_file_task(uploaded_file_id):
 def start_export_run(export_run_id, user):
     """Start the given export run."""
     exports.start_run(export_run_id, user)
+
+
+@task
+def shapefile_vacuum(directory):
+    """Put shapefile parts into zip files in directory."""
+    shapevac.shapefile_vacuum_directory(directory)